Workaround for Echidna.
Projects that don't use generate the init file in the buildscript won't
work without it since it's specified in the default config. Also it even
then it shouldn't be disabled since then the block gaslimit can no
longer be used to determine whether we're running in Echidna or Foundry.
We need to be able to do that because Echidna doesn't have the assume()
cheatcode and will just get stuck when trying to call it.